Fix/comments api (#43)
* Fix typo error

* Fix serialization format

* Add column validations

* Adjust the create&update action

* Remove boolean validation for anonymity column

* Rspec test for comments controller

+++ app/models/comment.rb
@@ -5,6 +5,9 @@ class Comment < ApplicationRecord
   belongs_to :user
   has_many :course_ratings, through: :user
 
+  validates :title, :content, presence: { message: '%{attribute} can not be empty' }
+  validates :course_id, presence: { message: 'Must specify a course' }
+
   # 重載course_ratings方法
   # 使其只返回該筆心得所對應到的評分紀錄
   def course_ratings
@@ -13,7 +16,7 @@ class Comment < ApplicationRecord
 
   def serializable_hash(options = nil)
     options = options.try(:dup) || {}
-    super({ **options, except: [:user_id, :course_id] }).tap do |result|
+    super({ **options, except: [:user_id, :course_id, :created_at, :updated_at] }).tap do |result|
       result[:course] = course.serializable_hash_for_comments
       result[:user] = { 'id': user_id, 'name': user.name }
       result[:rating] = '000'
@@ -40,10 +43,16 @@ class Comment < ApplicationRecord
 
   # 建立該筆心得對應的評分紀錄
   def create_course_ratings(ratings = [0, 0, 0])
-    ratings.each do |rating|
+    # return false if the ratings is nil
+    return false if ratings.nil?
+
+    ratings_array = ratings.scan(/\d/).map(&:to_i)
+
+    # Check if any rating is negative or larger than 5
+    ratings_array.each do |rating|
       return false if rating > 5 || rating.negative?
     end
-    ratings.each_with_index do |rating, index|
+    ratings_array.each_with_index do |rating, index|
       user.course_ratings.create course: course, category: index, score: rating
     end
     true
@@ -51,12 +60,17 @@ class Comment < ApplicationRecord
 
   # 更新該筆心得對應的評分紀錄
   def update_course_ratings(ratings = [0, 0, 0])
+    return if ratings.nil?
+
     previous_rating = course_ratings.order(:category).pluck(:score)
+    ratings_array = ratings.scan(/\d/).map(&:to_i)
+
+    # return if the ratings remain unchanged
     return if previous_rating.eql?(ratings) || ratings.nil?
 
     # Delete old ratings records
     course_ratings.delete_all
     # Create updated rating records
-    create_course_ratings(ratings)
+    create_course_ratings(ratings_array)
   end
 end
